implementation for once in a lifetime recipes, migrate all long values to double

+++ nearit/src/main/java/it/near/sdk/recipes/RecipeCooler.java
@@ -20,7 +20,7 @@ public class RecipeCooler {
 
     private static final String RECIPE_COOLER_PREFS_NAME = "NearRecipeCoolerPrefsName";
     private static final String LOG_MAP = "LOG_MAP";
-    static final long NEVER_REPEAT = -1L;
+    static final double NEVER_REPEAT = -1D;
     private static final String LATEST_LOG = "LATEST_LOG";
     static final String GLOBAL_COOLDOWN = "global_cooldown";
     static final String SELF_COOLDOWN = "self_cooldown";
@@ -60,7 +60,7 @@ public class RecipeCooler {
                 cooldown.get(GLOBAL_COOLDOWN) == null) return true;
 
         long expiredSeconds = (currentTime.currentTimestamp() - getLatestLogEntry()) / 1000;
-        return expiredSeconds >= (Long) cooldown.get(GLOBAL_COOLDOWN);
+        return expiredSeconds >= ((Double) cooldown.get(GLOBAL_COOLDOWN)).longValue();
     }
 
     private boolean selfCooldownCheck(Recipe recipe, Map<String, Object> cooldown) {
@@ -68,12 +68,12 @@ public class RecipeCooler {
                 cooldown.get(SELF_COOLDOWN) == null ||
                 !getRecipeLogMap().containsKey(recipe.getId())) return true;
 
-        if ((Long)cooldown.get(SELF_COOLDOWN) == NEVER_REPEAT &&
+        if ((Double)cooldown.get(SELF_COOLDOWN) == NEVER_REPEAT &&
                 getRecipeLogMap().containsKey(recipe.getId())) return false;
 
         long recipeLatestEntry = getRecipeLogMap().get(recipe.getId());
         long expiredSeconds = (currentTime.currentTimestamp() - recipeLatestEntry) / 1000;
-        return expiredSeconds >= (Long) cooldown.get(SELF_COOLDOWN);
+        return expiredSeconds >= ((Double) cooldown.get(SELF_COOLDOWN)).longValue();
     }
 
     /**
